% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-2019.R
\name{bs_build_2019}
\alias{bs_build_2019}
\title{Build real-time data from the 2019 deployment}
\usage{
bs_build_2019(out_dir = ".")
}
\arguments{
\item{out_dir}{The directory in which output files should be
generated.}
}
\value{
\code{out_dir}, invisibly.
}
\description{
Build real-time data from the 2019 deployment
}
\examples{
\dontrun{
bs_build_2019()
}

}
