% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-navigator.R
\name{bs_build_realtime_navigator}
\alias{bs_build_realtime_navigator}
\title{Build data for Navigator}
\usage{
bs_build_realtime_navigator(built_dir = ".", out_dir = build_dir)
}
\arguments{
\item{built_dir}{The output directory used for
\code{\link[=bs_build_realtime]{bs_build_realtime()}} that contains the built data files.}

\item{out_dir}{Where the NetCDF files should be created.}
}
\value{
\code{out_dir}, invisibly.
}
\description{
For the purposes of \href{https://navigator.oceansdata.ca/}{Navigator},
the BSRTO is treated as a single station at 74.605N/91.251W.
There is a single WMO code assigned to the station that is used for
the moorings that represent all three depths.
}
\examples{
\dontrun{
bs_build_realtime()
bs_build_realtime_navigator()
}
}
