% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs-cache.R
\name{bs_cache}
\alias{bs_cache}
\alias{bs_has_cache}
\title{Find a local cache of the BSRTO FTP server}
\usage{
bs_cache()

bs_has_cache()
}
\description{
Find a local cache of the BSRTO FTP server
}
\examples{
bs_cache()
bs_has_cache()

}
