% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs-example.R
\name{bs_example}
\alias{bs_example}
\title{Example files}
\usage{
bs_example(example_file)
}
\arguments{
\item{example_file}{An example file in the 'ex' directory of the 'bsrto'
package.}
}
\value{
An absolute path to the file
}
\description{
Example files
}
\examples{
bs_example("CTD_98911_10P_11_DN.ODF")

}
