% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{bs_flag}
\alias{bs_flag}
\alias{bs_flag_label}
\alias{bs_flag_scheme}
\title{Internal BSRTO flag scheme}
\usage{
bs_flag(flag_label)

bs_flag_label(flag)

bs_flag_scheme()
}
\arguments{
\item{flag_label}{A human-readable}

\item{flag}{An integer flag value}
}
\description{
Internal BSRTO flag scheme
}
\examples{
bs_flag(c("not assessed", "probably good data"))
bs_flag_label(1:5)
bs_flag_scheme()

}
