% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bs_ftp_snapshot_latest}
\alias{bs_ftp_snapshot_latest}
\title{Snapshot of the BSRTO FTP server}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 275993 rows and 3 columns.
}
\usage{
bs_ftp_snapshot_latest
}
\description{
Snapshot of the BSRTO FTP server
}
\keyword{datasets}
