% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-climate.R
\name{read_ec_climate_hourly}
\alias{read_ec_climate_hourly}
\alias{read_ec_climate_hourly_vector}
\title{Read Environment Canada Climate CSVs}
\usage{
read_ec_climate_hourly(file, pb = NULL, utc_offset = 0)

read_ec_climate_hourly_vector(file_vector, utc_offset = 0)
}
\arguments{
\item{file}{A file, URL, or connection. Files ending in .gz, .bz2, .xz, or
.zip will be automatically uncompressed; URLs will be automatically
downloaded. See \code{\link[readr:read_lines]{readr::read_lines()}} for a full description of how
this parameter is interpreted.}

\item{pb}{A \link[progress:progress_bar]{progress bar} or \code{NULL}.}

\item{utc_offset}{Environment Canada provides dates and times in
"local standard time", which has a fixed UTC offset (i.e., does not
observe daylight savings time).}

\item{file_vector}{A vector of files or URLs}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Read Environment Canada Climate CSVs
}
\examples{
ec_file <- bs_example("met/ec-climate_hourly_54199_2019_08.csv")
read_ec_climate_hourly(ec_file)
read_ec_climate_hourly_vector(ec_file)

}
