% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-icl.R
\name{read_icl}
\alias{read_icl}
\alias{read_icl_header}
\alias{read_icl_header_lines}
\title{Read icl files}
\usage{
read_icl(file, skip = length(read_icl_header_lines(file)) + 1, n_max = Inf)

read_icl_header(file, header_lines = read_icl_header_lines(file))

read_icl_header_lines(file, n_header = 100)
}
\arguments{
\item{file}{A file, URL, or connection. Files ending in .gz, .bz2, .xz, or
.zip will be automatically uncompressed; URLs will be automatically
downloaded. See \code{\link[readr:read_lines]{readr::read_lines()}} for a full description of how
this parameter is interpreted.}

\item{skip}{The number of header rows to skip}

\item{n_max}{Maximum number of rows to read.}

\item{header_lines}{A previously read value obtained from
\code{\link[=read_odf_header_lines]{read_odf_header_lines()}}.}

\item{n_header}{The starting guess for number of header lines.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Read icl files
}
\examples{
icl_file <- bs_example("icl/SAF2564_20191010_19.txt")
read_icl(icl_file)
read_icl_header(icl_file)
read_icl_header_lines(icl_file)

}
