% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-lgh.R
\name{read_lgh}
\alias{read_lgh}
\alias{read_lgh_vector}
\title{Read LGH log files}
\usage{
read_lgh(file, tz = "UTC", pb = NULL)

read_lgh_vector(file_vector, tz = "UTC")
}
\arguments{
\item{file}{A file, URL, or connection. Files ending in .gz, .bz2, .xz, or
.zip will be automatically uncompressed; URLs will be automatically
downloaded. See \code{\link[readr:read_lines]{readr::read_lines()}} for a full description of how
this parameter is interpreted.}

\item{tz}{Timezone for which date times are representative.}

\item{pb}{A \link[progress:progress_bar]{progress bar} or \code{NULL}.}

\item{file_vector}{A vector of files or URLs}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Read LGH log files
}
\examples{
lgh_file <- bs_example("lgh/20191010.lgH")
read_lgh(lgh_file)
read_lgh_vector(lgh_file)

library(tidyr)
read_lgh(lgh_file) \%>\% unnest(log_text)

}
