% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-pcm.R
\name{read_pcm}
\alias{read_pcm}
\alias{read_pcm_vector}
\title{Read PCM files}
\usage{
read_pcm(file, tz = "UTC", pb = NULL)

read_pcm_vector(file_vector, tz = "UTC")
}
\arguments{
\item{file}{A file, URL, or connection. Files ending in .gz, .bz2, .xz, or
.zip will be automatically uncompressed; URLs will be automatically
downloaded. See \code{\link[readr:read_lines]{readr::read_lines()}} for a full description of how
this parameter is interpreted.}

\item{tz}{Timezone for which date times are representative.}

\item{pb}{A \link[progress:progress_bar]{progress bar} or \code{NULL}.}

\item{file_vector}{A vector of files or URLs}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Read PCM files
}
\examples{
pcm_file <- bs_example("pcm/19101018.pcm")
read_pcm(pcm_file)
read_pcm_vector(pcm_file)


}
