% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-rdi.R
\name{read_rdi}
\alias{read_rdi}
\alias{read_rdi_vector}
\title{Read RDI files}
\usage{
read_rdi(file, tz = "UTC")

read_rdi_vector(file_vector, tz = "UTC")
}
\arguments{
\item{file}{A file, URL, or connection. Files ending in .gz, .bz2, .xz, or
.zip will be automatically uncompressed; URLs will be automatically
downloaded. See \code{\link[readr:read_lines]{readr::read_lines()}} for a full description of how
this parameter is interpreted.}

\item{tz}{Timezone for which date times are representative.}

\item{file_vector}{A vector of files or URLs}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Read RDI files
}
\examples{
rdi_file <- bs_example("rdi/19101018.rdi")
read_rdi(rdi_file)
read_rdi_vector(rdi_file)

}
