% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/science.R
\name{sea_level_pressure_from_barometric}
\alias{sea_level_pressure_from_barometric}
\alias{barometric_pressure_from_sea_level}
\title{Barometric/Sea-level pressure conversion}
\usage{
sea_level_pressure_from_barometric(
  p,
  h,
  g = 9.80665,
  M = 0.0289644,
  R0 = 8.31447,
  T0 = 288.15
)

barometric_pressure_from_sea_level(
  p0,
  h,
  g = 9.80665,
  M = 0.0289644,
  R0 = 8.31447,
  T0 = 288.15
)
}
\arguments{
\item{p}{Measured barometric pressure (can be any pressure unit)}

\item{h}{Elevation of barometric pressure measurement in meters}

\item{g}{Accelleration due to gravity in m/s2}

\item{M}{Molar mass of dry air (kg / mol)}

\item{R0}{Universal gas constant (mol-Kelvins)}

\item{T0}{Sea level standard temperature (Kelvins)}

\item{p0}{Pressure at sea level in (can be any pressure unit)}
}
\value{
The corrected pressure in the same unit as \code{p} or \code{p0}.
}
\description{
Formula from \url{https://en.wikipedia.org/wiki/Atmospheric_pressure}.
}
\examples{
sea_level_pressure_from_barometric(102113, 68)
barometric_pressure_from_sea_level(101300, 68)

}
