% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/science.R
\name{barrow_strait_declination}
\alias{barrow_strait_declination}
\title{Approximate declinations for Barrow Strait}
\usage{
barrow_strait_declination(date_time)
}
\arguments{
\item{date_time}{A POSIXct for which declenation should be calculated}
}
\value{
A numeric vector of approximated declinations
}
\description{
Approximate declinations for Barrow Strait
}
\examples{
# location of observatory
lon <- -91.25105
lat <- 74.60635

# from start of deployment to the the end of the model's validity
date_time <- seq(
  as.POSIXct("2019-01-01 00:00:00", tz = "UTC"),
  as.POSIXct("2025-12-31 11:59:00", tz = "UTC"),
  by = "day"
)
dec_model <- oce::magneticField(
  rep(lon, length(date_time)),
  rep(lat, length(date_time)),
  date_time
)
t <- as.numeric(date_time, origin = "1970-01-01 00:00:00")
lm_approx <- lm(dec_model$declination ~ poly(t, 2, raw = TRUE))

# take a look at the max error we can expect
max(residuals(lm_approx))

plot(t, dec_model$declination, type = "l")
lines(t, predict(lm_approx), col = "red")

# used to generate the values in this function
as.character(coef(lm_approx))

}
