% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs-cache.R
\name{bs_cache_dir}
\alias{bs_cache_dir}
\alias{bs_has_full_cache}
\title{Find a local cache of the BSRTO FTP server}
\usage{
bs_cache_dir(...)

bs_has_full_cache()
}
\arguments{
\item{...}{Passed to \code{\link[=file.path]{file.path()}}}
}
\description{
Find a local cache of the BSRTO FTP server
}
\examples{
bs_cache_dir()
bs_has_full_cache()

}
