% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs-ftp.R
\name{bs_cached}
\alias{bs_cached}
\alias{bs_ftp_list}
\title{Cache, fetch, and list files from the BSRTO FTP server}
\usage{
bs_cached(
  x,
  ftp = bs_ftp_server(),
  cache = bs_cache_dir(),
  async = FALSE,
  retries = 4,
  quiet = FALSE
)

bs_ftp_list(
  x,
  pattern = NULL,
  recursive = FALSE,
  ftp = bs_ftp_server(),
  print = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{x}{A path relative to /pub/dfo}

\item{ftp}{The location of the FTP server}

\item{cache}{Path to a local copy of /pub/dfo where files will be cached
or written.}

\item{async}{Use \code{TRUE} to download files asynchronously (much faster
for many small files)}

\item{retries}{The number of retries to use}

\item{quiet}{Use \code{TRUE} to suppress progress messages}

\item{pattern}{A regular expression against which files are matched
or \code{NULL} to match all files. See \code{\link[stringr:str_detect]{stringr::str_detect()}} for how
this argument is interpreted.}

\item{recursive}{Use \code{TRUE} to recurse directories}

\item{print}{Use \code{TRUE} to print out matching files as the listings are
downloaded.}
}
\value{
\code{bs_cached()} returns the file path to the local version of the
file after downloading; \code{bs_ftp_list()} returns a list of files (not
directories!) that match \code{pattern}.
}
\description{
Cache, fetch, and list files from the BSRTO FTP server
}
\examples{
\dontrun{
bs_cached("barrow/BarrowStraitDataSummary.xlsx")
bs_ftp_list("barrow")
}
}
