% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-hpb.R
\name{read_hpb}
\alias{read_hpb}
\title{Read hpb files}
\usage{
read_hpb(file_vector)
}
\arguments{
\item{file_vector}{A vector of filenames}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Read hpb files
}
\examples{
hpb_files <- list.files(bs_example("hpb"), "\\\\.hpb$", full.names = TRUE)
read_hpb(hpb_files)

}
