% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-ips.R
\name{read_ips_bn}
\alias{read_ips_bn}
\title{Read IPS files}
\usage{
read_ips_bn(file_vector)
}
\arguments{
\item{file_vector}{One ore more paths to local or remote .bnX files.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Read IPS files
}
\examples{
bn_file <- bs_example("ips/191010AA.bn1")
read_ips_bn(bn_file)

}
